software/python/pyproject.toml
[tool.poetry]
name = "wfmi"
version = "0.1.0"
description = "Weighted and Filtered Mutual Information Algorithm"
authors = [
    "Tommy P. Keane <talk.to.tommypkeane@gmail.com>"
]
maintainers = [
    "Tommy P. Keane <talk.to.tommypkeane@gmail.com>",
]
homepage = "https://www.tommypkeane.com"
repository = "https://github.com/tommypkeane/research-msc"
license = "CC-BY-NC-ND-4.0"
readme = "README.md"
packages = [
    { include = "wfmi" },
]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.2"
pillow = "^9.4.0"
opencv-python-headless = "^4.7.0.72"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
console_output_style = "percent"
log_auto_indent = true
log_level = "DEBUG"
log_cli = false  # Set true if you wanna see the logs live, otherwise look at `log_file`
log_cli_level = "DEBUG"
log_cli_date_format = "%Y-%m-%dT%H:%M:%S%z (%Z)"
log_cli_format = "%(asctime)s :: %(name)s :: %(levelname)s :: %(message)s"
addopts = "--cov"
testpaths = [
    "tests_unit",
]
